backend/requirements.txt
fastapi
uvicorn
requests
trafilatura
python-dotenv
qdrant-client
cohere
pydantic
